# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Uganda Diaspora Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string } }
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # NEWS
...
  /tourism:
    get:
      operationId: listTourism
      tags: [tourism]
      summary: List tourism attractions
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Tourism list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourismListResponse"
    post:
      operationId: createTourism
      tags: [tourism]
      summary: Create tourism attraction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourismInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourismAttraction"
  /tourism/{id}:
    get:
      operationId: getTourism
      tags: [tourism]
      summary: Get tourism attraction
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Attraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourismAttraction"
        "404":
          description: Not found
    patch:
      operationId: updateTourism
      tags: [tourism]
      summary: Update tourism attraction
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourismUpdate"
      responses:
        "200":
          description: Updated attraction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourismAttraction"
    delete:
      operationId: deleteTourism
      tags: [tourism]
      summary: Delete tourism attraction
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # WEBINARS
  /webinars:
    get:
      operationId: listWebinars
      tags: [webinars]
      summary: List webinars
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: category, in: query, schema: { type: string } }
        - { name: upcoming, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Webinar list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebinarListResponse"
    post:
      operationId: createWebinar
      tags: [webinars]
      summary: Create webinar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebinarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webinar"
  /webinars/{id}:
    get:
      operationId: getWebinar
      tags: [webinars]
      summary: Get webinar
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Webinar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webinar"
        "404":
          description: Not found
    patch:
      operationId: updateWebinar
      tags: [webinars]
      summary: Update webinar
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebinarUpdate"
      responses:
        "200":
          description: Updated webinar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webinar"
    delete:
      operationId: deleteWebinar
      tags: [webinars]
      summary: Delete webinar
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # EVENTS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, role, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
        country: { type: ["string", "null"] }
        profession: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserUpdate:
      type: object
      properties:
        fullName: { type: string }
        country: { type: string }
        profession: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        role: { type: string }
        isActive: { type: boolean }
    UserListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # NEWS
...
    TourismAttraction:
      type: object
      required: [id, name, category, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        location: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        gallery: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        entryFee: { type: ["string", "null"] }
        openingHours: { type: ["string", "null"] }
        contactPhone: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        createdAt: { type: string, format: date-time }
    TourismInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        location: { type: string }
        imageUrl: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        entryFee: { type: string }
        openingHours: { type: string }
        contactPhone: { type: string }
        website: { type: string }
        isFeatured: { type: boolean }
    TourismUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        location: { type: string }
        imageUrl: { type: string }
        isFeatured: { type: boolean }
    TourismListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TourismAttraction"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # WEBINAR
    Webinar:
      type: object
      required: [id, title, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        youtubeUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        speakerName: { type: ["string", "null"] }
        scheduledAt: { type: ["string", "null"] }
        isLive: { type: boolean }
        isPublished: { type: boolean }
        viewCount: { type: integer }
        registrationCount: { type: integer }
        createdAt: { type: string, format: date-time }
    WebinarInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        youtubeUrl: { type: string }
        thumbnailUrl: { type: string }
        category: { type: string }
        speakerName: { type: string }
        scheduledAt: { type: string }
        isLive: { type: boolean }
        isPublished: { type: boolean }
    WebinarUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        youtubeUrl: { type: string }
        thumbnailUrl: { type: string }
        category: { type: string }
        speakerName: { type: string }
        scheduledAt: { type: string }
        isLive: { type: boolean }
        isPublished: { type: boolean }
    WebinarListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Webinar"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # EVENTS
...