# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Uganda Diaspora Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  fullName?: string;
  country?: string;
  profession?: string;
  bio?: string;
  avatarUrl?: string;
  role?: string;
  isActive?: boolean;
}